+++ src/hud/waiticondisplay.cpp
@@ -45,13 +45,18 @@ void Render(glm::vec2 pos, glm::vec4 opacityTint, DialoguePageMode mode) {
       glm::vec4 col = glm::vec4(1.0f, 1.0f, 1.0f, opacityTint.w);
       // Erin DialogueBox
       if (mode == DPM_REV) {
-        Renderer->DrawSprite(SpriteAnim.CurrentSprite(),
-            glm::vec2(pos.x + Impacto::Profile::CHLCC::DialogueBox::REVWaitIconOffset.x,
-                      pos.y + Impacto::Profile::CHLCC::DialogueBox::REVWaitIconOffset.y),
+        Renderer->DrawSprite(
+            SpriteAnim.CurrentSprite(),
+            glm::vec2(
+                pos.x +
+                    Impacto::Profile::CHLCC::DialogueBox::REVWaitIconOffset.x,
+                pos.y +
+                    Impacto::Profile::CHLCC::DialogueBox::REVWaitIconOffset.y),
             col);
-      
+
       } else {
-        Renderer->DrawSprite(SpriteAnim.CurrentSprite(),
+        Renderer->DrawSprite(
+            SpriteAnim.CurrentSprite(),
             glm::vec2(pos.x + WaitIconOffset.x, pos.y + WaitIconOffset.y), col);
       }
     } else {
@@ -66,8 +71,8 @@ void Render(glm::vec2 pos, glm::vec4 opacityTint, DialoguePageMode mode) {
         SpriteAnim.CurrentSprite(),
         glm::vec2(WaitIconOffset.x - 50, WaitIconOffset.y - 50), opacityTint);*/
     Renderer->DrawSprite(WaitIconSprite,
-                           glm::vec2(WaitIconOffset.x, WaitIconOffset.y),
-                           opacityTint, glm::vec2(1.0f));
+                         glm::vec2(WaitIconOffset.x, WaitIconOffset.y),
+                         opacityTint, glm::vec2(1.0f));
   } else if (WaitIconCurrentType == +WaitIconType::RotateZ) {
     // TODO: MO6TW only for now
     glm::vec3 euler(SimpleAnim.Progress * 2.0f * M_PI, 0, 0.6f);
@@ -91,7 +96,7 @@ void Render(glm::vec2 pos, glm::vec4 opacityTint, DialoguePageMode mode) {
     Renderer->DrawSprite(
         WaitIconSprite,
         glm::vec2(pos.x + WaitIconOffset.x, pos.y + WaitIconOffset.y),
-        opacityTint, glm::vec2(1.0f), SimpleAnim.Progress * 2.0f * M_PI);
+        opacityTint, glm::vec2(1.0f), SimpleAnim.Progress * 2.0f * (float)M_PI);
   }
 }
