+++ src/renderer/dx9/yuvframe.h
@@ -13,7 +13,7 @@ class DX9YUVFrame : public YUVFrame {
  public:
   DX9YUVFrame(IDirect3DDevice9* device);
 
-  void Init(int width, int height) override;
+  void Init(float width, float height) override;
 
   void Submit(void* luma, void* cb, void* cr) override;
   void Release() override;
